BUILD: quic: fix wrong potential NULL dereference
GCC warns about a possible NULL dereference when requeuing a datagram on
the connection socket. This happens due to a MT_LIST_POP to retrieve a
rxbuf instance.

In fact, this can never be NULL there is enough rxbuf allocated for each
thread. Once a thread has finished to work with it, it must always
reappend it.

This issue was introduced with the following patch :
  commit b34d353968db7f646e83871cb6b21a246af84ddc
  BUG/MEDIUM: quic: consume contig space on requeue datagram
As such, it must be backported in every version with the above commit.

This should fix the github CI compilation error.

(cherry picked from commit f40a72a7fffe5a926d8a01547fa19a686fe591b5)
Signed-off-by: Amaury Denoyelle <adenoyelle@haproxy.com>