BUG/MINOR: quic: allow-0rtt warning must only be emitted with quic bind
When built with USE_QUIC_OPENSSL_COMPAT, a warning is emitted when using
allow-0rtt. However this warning is emitted for every allow-0rtt
keywords on the bind line which is confusing, it must only be done in
case the bind is a quic one. Also this does not handle the case where
the allow-0rtt keyword is in the crt-list.

This patch moves the warning to ssl_quic_initial_ctx() in order to emit
the warning in every useful cases.

(cherry picked from commit 8c004153e5f1802b2060629d8656b42b84c65bbd)
[wla: quic_ssl.c does not exist in 2.8]
Signed-off-by: William Lallemand <wlallemand@haproxy.com>