MINOR: map: accelerate "set-map" action.
Add an eb_root member to the map (pat_ref struct) and an ebpt_node to its element
(pat_ref_elt struct) and modify the code to insert these nodes to their ebtree
each time they are allocated. This is done in pat_ref_append().

This patch modify http_action_set_map() which is the callback at least used
by "set-map" action. The pat_ref_elt element returned by pat_ref_find_elt() is
no more ignored, but reused if not NULL by pat_ref_set() as first element to
lookup from. This latter is also modified to use the ebtree attached to the map
in place of the ->head list attached to each map element (pat_ref_elt struct).

Also modifies pat_ref_find_elt() to makes it use the map ebtre to the map
in place of inspecting all the element with a strcmp() call.