BUILD: quic: Compilation issue on 32-bits systems with quic_may_send_bytes()
quic_may_send_bytes() implementation arrived with this commit:

  MINOR: quic: Amplification limit handling sanitization.

It returns a size_t. So when compared with QUIC_MIN() with qc->path->mtu there is
no need to cast this latter anymore because it is also a size_t.

Detected when compiled with -m32 gcc option.

Thank you to @chipitsine for having reported this issue on GH #2261