DOC: lua: fix core.register_action typo
"converter" was used in place of "action" as a result of a copy-paste
error probably.

Also, rephrasing the "actions" keyword explanation to prevent confusion
between action name (which is the name of new action about to be created)
and action facilities where we want to expose the new action.

This could be backported to every stable versions.

(cherry picked from commit 7e9aea789f4412bbcb149f9954b08da861c2fc7e)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>