BUG/MINOR: quic: Possible skipped RTT sampling
There are very few chances this bug may occur. Furthermore the consequences
are not dramatic: an RTT sampling may be ignored. I guess this may happen
when the now_ms global value wraps.

Do not rely on the time variable value a packet was sent to decide if it
is a newly acknowledged packet but on its presence or not in the tx packet
ebtree.

Must be backported as far as 2.6.

(cherry picked from commit 3c90c1ce6b42ab53553ff06fc83781d4327825dc)
[cF: applied on quic_conn.c because qui_rx.c does not exist]
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>