BUG/MEDIUM: stconn: Don't block sends if there is a pending shutdown
For the same reason than the previous patch, we must not block the sends
when there is a pending shutdown. In other words, we must consider the sends
are allowed when there is a pending shutdown.

This patch must slowly be backported as far as 2.2. It should partially fix
issue #2249.

(cherry picked from commit b50a471adbb464179a040ec9488dda86da38a07e)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>