BUG/MEDIUM: stconn: Report read activity when a stream is attached to front SC
It only concerns the front SC. But it is important to report a read activity
when a stream is created and attached to the front SC, especially in TCP. In
HTTP, when this happens, the request was necessarily received. But in TCP,
the client may open a connection without sending anything. We must still
report a first read activity in this case to be able to properly report
client timeout.

This patch must be backported to 2.8.

(cherry picked from commit 34645a6365827f1cd3929afe8c8e7055fa4556ee)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>