BUG/MINOR: ring/cli: Don't expect input data when showing events
The "show events" command may wait for now events if "-w" option is used. In
this case, no timeout must be triggered. So we explicitly state no input
data are expected. This disables the read timeout on the client side.

This patch should be backported to 2.8. It is probably useless to backport
it further. In all cases, it depends on the commit "BUG/MINOR: applet:
Always expect data when CLI is waiting for a new command"

(cherry picked from commit 700ca14fc1ff14ddc573620940cf0ac7872c7ab0)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>