BUG/MINOR: quic: Leak of frames to send.
In very rare cases, it is possible that packet are detected as lost, their frames
requeued, then the connection is released without releasing for any reason (to
be killed because of a sendto() fatal failure for instance. Such frames are lost
and never release because the function which release their packet number spaces
does not release the frames which are still enqueued to be send.

Must be backported as far as 2.6.

(cherry picked from commit 3921bf80c7c2933afbb93ea595b1797bcfc44982)
[wt: bkp from Fred: changes goes to quic_conn_release() in 2.8]
Signed-off-by: Willy Tarreau <w@1wt.eu>