BUG/MINOR: mux-h1: Ignore C-L when sending H1 messages if T-E is also set
In fact, it is already done but both flags (H1_MF_CLEN and H1_MF_CHUNK) are
set on the H1 parser. Thus it is errorprone when H1 messages are sent,
especially because most of time, the "Content-length" case is processed
before the "chunked" one. This may lead to compute the wrong chunk size and
to miss the last chunk.

This patch must be backported as far as 2.6. This case is not handled in 2.4
and lower.

(cherry picked from commit c367957851b17f1c153c4d0f75add35dea957d51)
 [ad: h1_process_mux() has been splitted in 2.9. Move statement from
  h1_make_eoh() to h1_process_mux()]
Signed-off-by: Amaury Denoyelle <adenoyelle@haproxy.com>