MINOR: hlua: Save the lua socket's timeout in its context
When the lua socket timeout is set, it is now saved in its context. If there
is already a stream attached to the appctx, the timeout is then immediately
modified. Otherwise, it is modified when the stream is created, thus during
the appctx initialization.

For now, the appctx is initialized when it is created. But this will change
to fix issues with the lua sockets. Thus, this patch is mandatory.

(cherry picked from commit 0be1ae2fa2d04584027fa09af2e3b54c0a0786b7)
Signed-off-by: Amaury Denoyelle <adenoyelle@haproxy.com>